Edit download link
Currently, it suggests finding th zip file link in the dashboard. However, since the GitHub redesign, this link isnt there anymore - it is in a dropdown at the top. I have just linked to the file and suggested using git for those who can. I think this is a bit easier.